+++ packages/f2elint/src/actions/scan.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import fs from 'fs-extra';
-import execa from 'execa';
 import glob from 'glob';
+import prettier from 'prettier';
 import * as eslint from '../lints/eslint';
 import * as stylelint from '../lints/stylelint';
 import * as markdownLint from '../lints/markdownlint';
@@ -11,6 +11,8 @@ import {
   ESLINT_FILE_EXT,
   STYLELINT_FILE_EXT,
   MARKDOWN_LINT_FILE_EXT,
+  PRETTIER_FILE_EXT,
+  PRETTIER_IGNORE_PATTERN,
 } from '../utils/constants';
 import type { ScanOptions, ScanResult, PKG, Config, ScanReport } from '../types';
 
@@ -30,6 +32,23 @@ export default async (options: ScanOptions): Promise<ScanReport> => {
   const config: Config = readConfigFile(`${PKG_NAME}.config.js`);
   const runErrors: Error[] = [];
   let results: ScanResult[] = [];
+
+  // prettier
+  if (fix && config.enablePrettier !== false) {
+    const files = options.files
+      ? (getLintFiles(PRETTIER_FILE_EXT) as string[])
+      : glob.sync(`**/*.{${PRETTIER_FILE_EXT.map((t) => t.replace(/^\./, '')).join(',')}}`, {
+          cwd,
+          ignore: PRETTIER_IGNORE_PATTERN,
+        });
+    files.forEach((filepath) => {
+      const text = fs.readFileSync(filepath, 'utf8');
+      prettier.resolveConfig(filepath).then((options) => {
+        const formatted = prettier.format(text, { ...options, filepath });
+        fs.writeFileSync(filepath, formatted, 'utf8');
+      });
+    });
+  }
 
   // eslint
   try {
@@ -67,22 +86,6 @@ export default async (options: ScanOptions): Promise<ScanReport> => {
         files,
       });
       results = results.concat(markdownLint.formatResults(data, quiet));
-    } catch (e) {
-      runErrors.push(e);
-    }
-  }
-
-  // prettier 格式化
-  if (config.enablePrettier && options.fix) {
-    try {
-      const ext = [...ESLINT_FILE_EXT, ...STYLELINT_FILE_EXT];
-      const files = options.files
-        ? getLintFiles(ext)
-        : glob.sync(`**/*.{${ext.map((t) => t.replace(/^\./, '')).join(',')}}`, {
-            cwd,
-            ignore: 'node_modules/**',
-          });
-      execa.sync(path.resolve(__dirname, '../../node_modules/.bin/prettier'), ['-w', ...files]);
     } catch (e) {
       runErrors.push(e);
     }