+++ packages/f2elint/src/utils/constants.ts
@@ -84,6 +84,9 @@ export const ESLINT_IGNORE_PATTERN: string[] = [
   'node_modules/',
   'build/',
   'dist/',
+  'coverage/',
+  'es/',
+  'lib/',
   '**/*.min.js',
   '**/*-min.js',
   '**/*.bundle.js',
@@ -101,6 +104,9 @@ export const STYLELINT_IGNORE_PATTERN: string[] = [
   'node_modules/',
   'build/',
   'dist/',
+  'coverage/',
+  'es/',
+  'lib/',
   '**/*.min.css',
   '**/*-min.css',
   '**/*.bundle.css',
@@ -114,4 +120,32 @@ export const MARKDOWN_LINT_FILE_EXT: string[] = ['.md'];
 /**
  * markdownLint 扫描忽略的文件或文件目录
  */
-export const MARKDOWN_LINT_IGNORE_PATTERN: string[] = ['node_modules/', 'build/', 'dist/'];
+export const MARKDOWN_LINT_IGNORE_PATTERN: string[] = [
+  'node_modules/',
+  'build/',
+  'dist/',
+  'coverage/',
+  'es/',
+  'lib/',
+];
+
+/**
+ * Prettier 扫描文件扩展名
+ */
+export const PRETTIER_FILE_EXT = [
+  ...STYLELINT_FILE_EXT,
+  ...ESLINT_FILE_EXT,
+  ...MARKDOWN_LINT_FILE_EXT,
+];
+
+/**
+ * Prettier 扫描忽略的文件或文件目录
+ */
+export const PRETTIER_IGNORE_PATTERN: string[] = [
+  'node_modules/**/*',
+  'build/**/*',
+  'dist/**/*',
+  'lib/**/*',
+  'es/**/*',
+  'coverage/**/*',
+];